bot/requirements.txt
-i https://pypi.org/simple
anyio==4.7.0; python_version >= '3.9'
asttokens==3.0.0; python_version >= '3.8'
attrs==24.3.0; python_version >= '3.8'
certifi==2024.12.14; python_version >= '3.6'
cffi==1.17.1; os_name == 'nt' and implementation_name != 'pypy'
charset-normalizer==3.4.1; python_version >= '3.7'
colorama==0.4.6; python_version >= '2.7' and python_version not in '3.0, 3.1, 3.2, 3.3, 3.4, 3.5, 3.6'
et-xmlfile==2.0.0; python_version >= '3.8'
execnet==2.1.1; python_version >= '3.8'
executing==2.1.0; python_version >= '3.8'
greenlet==3.1.1; python_version < '3.13' and platform_machine == 'aarch64' or (platform_machine == 'ppc64le' or (platform_machine == 'x86_64' or (platform_machine == 'amd64' or (platform_machine == 'AMD64' or (platform_machine == 'win32' or platform_machine == 'WIN32')))))
h11==0.14.0; python_version >= '3.7'
httpcore==1.0.7; python_version >= '3.8'
httpx==0.28.1; python_version >= '3.8'
icecream==2.1.3
idna==3.10; python_version >= '3.6'
iniconfig==2.0.0; python_version >= '3.7'
mysql-connector-python==9.1.0; python_version >= '3.9'
numpy==2.2.1; python_version == '3.11'
openpyxl==3.1.5; python_version >= '3.8'
outcome==1.3.0.post0; python_version >= '3.7'
packaging==24.2; python_version >= '3.8'
pandas==2.2.3; python_version >= '3.9'
pluggy==1.5.0; python_version >= '3.8'
psycopg2==2.9.10; python_version >= '3.8'
pycparser==2.22; python_version >= '3.8'
pygments==2.18.0; python_version >= '3.8'
pysocks==1.7.1
pytest==8.3.4; python_version >= '3.8'
pytest-shutil==1.8.1; python_version >= '3.6'
python-dateutil==2.9.0.post0; python_version >= '2.7' and python_version not in '3.0, 3.1, 3.2, 3.3'
python-dotenv==1.0.1; python_version >= '3.8'
python-json-logger==3.2.1; python_version >= '3.8'
python-telegram-bot==21.10; python_version >= '3.9'
pytz==2024.2
requests==2.32.3; python_version >= '3.8'
selenium==4.27.1; python_version >= '3.8'
six==1.17.0; python_version >= '2.7' and python_version not in '3.0, 3.1, 3.2, 3.3'
sniffio==1.3.1; python_version >= '3.7'
sortedcontainers==2.4.0
sqlalchemy==2.0.36; python_version >= '3.7'
telegram==0.0.1
termcolor==2.5.0; python_version >= '3.9'
trio==0.28.0; python_version >= '3.9'
trio-websocket==0.11.1; python_version >= '3.7'
typing-extensions==4.12.2; python_version >= '3.8'
tzdata==2024.2; python_version >= '2'
urllib3[socks]==2.3.0; python_version >= '3.9'
websocket-client==1.8.0; python_version >= '3.8'
wsproto==1.2.0; python_full_version >= '3.7.0'
